requirements.txt
Flask==2.3.2
python-docx==0.8.11
gunicorn==21.2.0
python-dotenv==1.0.0